Zoomed Axes and Widget Interaction Improved
Widget which were drawn in an unzoomed axes could not been moved, when
one zoomed in and part of the widget is outside the image. This is now
possible. The axes size is globally stored in axes_xlim/ axes_ylim only
updated, if a larger axes is recgonized. Which means, movement of
widgets outside of the zoomed axes is possible if a widget was drawn
inside of an unzoomed axes.