Bugfix: Issue with static border variables
Clear Widget did not reset the BorderLimits to Default value. Added a
static method to the DraggableWidgets to the the borderlimits to None.
Important if there is a change of axis limits for example by changing
images.

Also get_widget now returns a None if there is not yet a line in the
drawn_line/ drawn_rectangle lists. This is confrom with PEP8